No longer strongly sign assemblies; Correct post-build event invocation
I attempted to mock Unity tests within VS's TestRunner, and discovered that
Unity does not strongly sign their assemblies, which posed multiple levels
of challenges to that goal. So for now, just to be consistent with them,
I'm going to leave these unsigned. If this needs to change in the future,
it's easier to sign it then than to unsign it if already signed.